Revert "refactor: extract Pitch component to separate files"
This reverts commit f0eca90bc16011fbe6094d7cfefeb1a367a262cf.
