--- src/components/Pitch/IPitch.ts
@@ -1,5 +1,0 @@
-import {ComponentProps} from 'react'
-
-export interface IPitch extends ComponentProps<'aside'> {
-    srcPath: string
-}
