--- src/components/Pitch/index.ts
@@ -1,1 +1,0 @@
-export * from './Pitch'
