+++ src/components/index.ts
@@ -10,4 +10,3 @@ export * from './SkillModal'
 export * from './SkillCard'
 export * from './ProjectCard'
 export * from './ProjectFilter'
-export * from './Pitch'
