+++ src/pages/About/About.tsx
@@ -1,14 +1,14 @@
-import {SocialLink} from '../../types/SocialLink'
-import {ComponentProps, PropsWithChildren} from 'react'
 import {faUpRightFromSquare} from '@fortawesome/free-solid-svg-icons'
 
-import {Navbar, Pitch} from '../../components'
+import {Navbar} from '../../components'
 
 import './About.css'
+import {ComponentProps, PropsWithChildren} from 'react'
+import {SocialLink} from '../../types/SocialLink'
 
 function Root(props: PropsWithChildren) {
     const {children} = props
-
+    
     return (
         <article className='about' id='about'>
             {children}
@@ -22,10 +22,10 @@ interface IHeader extends ComponentProps<'header'> {
 
 function Header(props: IHeader) {
     const {links} = props
-
+    
     return (
         <header className='about__header'>
-            <h2 className='about__title'>
+            <h2 className='about__title' >
                 sobre mim
             </h2>
 
@@ -40,6 +40,26 @@ function Header(props: IHeader) {
     )
 }
 
+interface IPitch extends ComponentProps<'aside'> {
+    srcPath: string
+}
+
+function Pitch(props: IPitch) {
+    const {srcPath} = props
+    
+    return (
+        <aside className='window-frame pitch'>
+            <video className='pitch__video' src={srcPath}></video>
+
+            <menu className='pitch__controls'>
+                <li className='pitch__control'>
+                    <button className='pitch__button'>play</button>
+                </li>
+            </menu>
+        </aside>
+    )
+}
+
 interface IAbout extends ComponentProps<'article'> {
     links: SocialLink[]
     srcPath: string
@@ -47,7 +67,7 @@ interface IAbout extends ComponentProps<'article'> {
 
 function About(props: IAbout) {
     const {links, srcPath} = props
-
+    
     return (
         <Root>
             <Header links={links}/>
